Prefer Framework::Definition::Base to Framework::Definition
Avoids the slightly circular nature of putting a definition in a
Definition namespace when that namespace is also a class.

Framework is still a class, though.
